+++ README.md
@@ -41,7 +41,7 @@ Start with the [Hello World](basics/hello-world) and then jump to the topic you 
 - [Jupyter Widgets](widget-tracker/jupyter-widgets)
 - [Kernel Output](advanced/kernel-output)
 - [Kernel Messaging](advanced/kernel-messaging)
-- [Server Hello World](advanced/myserverextension)
+- [Server Hello World](advanced/server-extension)
 - [Serving Files](advanced/serving-files)
 
 You can expect from each example:
@@ -106,12 +106,12 @@ We have structured the examples based on the [extension points](https://jupyterl
 
 ## Advanced
 
-| Example                                          | Preview                                                                                     |
-| ------------------------------------------------ | ------------------------------------------------------------------------------------------- |
-| [Kernel Output](advanced/kernel-output)          | [![OutputArea class](advanced/kernel-output/preview.gif)](advanced/kernel-output)           |
-| [Kernel Messaging](advanced/kernel-messaging)    | [![Kernel Messages](advanced/kernel-messaging/preview.gif)](advanced/kernel-messaging)      |
-| [Server Hello World](advanced/myserverextension) | [![Server Hello World](advanced/myserverextension/preview.png)](advanced/myserverextension) |
-| [Serving Files](advanced/serving-files)          |                                                                                             |
+| Example                                         | Preview                                                                                   |
+| ----------------------------------------------- | ----------------------------------------------------------------------------------------- |
+| [Kernel Output](advanced/kernel-output)         | [![OutputArea class](advanced/kernel-output/preview.gif)](advanced/kernel-output)         |
+| [Kernel Messaging](advanced/kernel-messaging)   | [![Kernel Messages](advanced/kernel-messaging/preview.gif)](advanced/kernel-messaging)    |
+| [Server Hello World](advanced/server-extension) | [![Server Hello World](advanced/server-extension/preview.png)](advanced/server-extension) |
+| [Serving Files](advanced/serving-files)         |                                                                                           |
 
 ## Prerequisites
