+++ advanced/server-extension/MANIFEST.in
@@ -2,11 +2,11 @@ include LICENSE
 include README.md
 
 include setupbase.py
-include jupyter-config/myserverextension.json
+include jupyter-config/server_extension.json
 
 include package.json
 include ts*.json
-include myserverextension/labextension/*.tgz
+include server_extension/labextension/*.tgz
 
 # Javascript files
 graft src
