+++ advanced/server-extension/jupyter-config/server_extension.json
@@ -1,7 +1,7 @@
 {
   "NotebookApp": {
     "nbserver_extensions": {
-      "myserverextension": true
+      "server_extension": true
     }
   }
 }
